# api/openapi.yaml
openapi: 3.0.3
info:
  title: Opero API
  version: 0.1.0
  description: >
    Opero people-operations API. This spec is the single source of truth —
    every operation is defined here before it is implemented. Operational
    routes like /health are infrastructure and intentionally not part of the
    contract.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user.
      description: >
        Returns the authenticated user and their tenant, resolved from the
        bearer token. Requires authentication.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The current user and tenant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records.
      description: Lists attendance records, optionally filtered by employee, status, and a [from, to) window on check-in time.
      security:
        - bearerAuth: []
      parameters:
        - name: employee_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [checked_in, checked_out, missed]
        - name: from
          in: query
          required: false
          description: Inclusive lower bound on check_in_at (ISO-8601).
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Exclusive upper bound on check_in_at (ISO-8601).
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Matching attendance records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttendanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in (field staff).
      description: >
        Records a check-in for the authenticated employee (resolved from the
        token). Idempotent on client_id: replaying the same client_id returns
        the existing record rather than creating a duplicate — this is what
        makes the offline mobile queue safe to replay.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInRequest'
      responses:
        '200':
          description: Existing record returned (idempotent replay).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '201':
          description: Checked in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: The client_id is already in use by another employee.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /attendance/check-out:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Check out (field staff).
      description: >
        Records check-out against an existing attendance record identified by
        the same client_id used at check-in. Idempotent.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckOutRequest'
      responses:
        '200':
          description: Checked out (or already checked out — idempotent).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: A stable, machine-readable error code.
          example: invalid_credentials
        message:
          type: string
          description: A human-readable explanation.
...
    CurrentUserResponse:
      type: object
      required: [user, tenant]
      properties:
        user:
          $ref: '#/components/schemas/UserSummary'
        tenant:
          $ref: '#/components/schemas/TenantSummary'
    UserSummary:
      type: object
      required: [id, email, role, status]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, manager, employee]
        status:
          type: string
          enum: [active, disabled]
    TenantSummary:
      type: object
      required: [id, name, slug, status, plan]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, suspended, provisioning]
        plan:
          type: string
...
    AttendanceRecord:
      type: object
      required: [id, employee_id, client_id, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        shift_id:
          type: string
          format: uuid
          nullable: true
        client_id:
          type: string
          format: uuid
          description: Client-generated id; the idempotency key for check-in/out.
        check_in_at:
          type: string
          format: date-time
          nullable: true
        check_in_lat:
          type: number
          format: double
          nullable: true
        check_in_lng:
          type: number
          format: double
          nullable: true
        check_in_photo_url:
          type: string
          nullable: true
        check_out_at:
          type: string
          format: date-time
          nullable: true
        check_out_lat:
          type: number
          format: double
          nullable: true
        check_out_lng:
          type: number
          format: double
          nullable: true
        check_out_photo_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [checked_in, checked_out, missed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CheckInRequest:
      type: object
      required: [client_id]
      properties:
        client_id:
          type: string
          format: uuid
          description: Client-generated idempotency key (stable across offline retries).
        shift_id:
          type: string
          format: uuid
          nullable: true
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
        photo_url:
          type: string
          nullable: true
    CheckOutRequest:
      type: object
      required: [client_id]
      properties:
        client_id:
          type: string
          format: uuid
          description: The same client_id used at check-in.
        lat:
          type: number
          format: double
          nullable: true
        lng:
          type: number
          format: double
          nullable: true
        photo_url:
          type: string
          nullable: true
...