# backend/gen/sqlc/identity/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package identitydb

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type AttendanceRecord struct {
	ID               uuid.UUID
	EmployeeID       uuid.UUID
	ShiftID          pgtype.UUID
	ClientID         uuid.UUID
	CheckInAt        pgtype.Timestamptz
	CheckInLat       *float64
	CheckInLng       *float64
	CheckInPhotoUrl  *string
	CheckOutAt       pgtype.Timestamptz
	CheckOutLat      *float64
	CheckOutLng      *float64
	CheckOutPhotoUrl *string
	Status           string
	CreatedAt        time.Time
	UpdatedAt        time.Time
}

type Department struct {
	ID        uuid.UUID
	Name      string
	ParentID  pgtype.UUID
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Employee struct {
	ID             uuid.UUID
	UserID         pgtype.UUID
	FullName       string
	Email          *string
	Phone          *string
	EmploymentType string
	DepartmentID   pgtype.UUID
	Title          *string
	Status         string
	HiredAt        pgtype.Date
	CreatedAt      time.Time
	UpdatedAt      time.Time
	RoleID         pgtype.UUID
}

type Location struct {
	ID        uuid.UUID
	Name      string
	Address   *string
	Lat       *float64
	Lng       *float64
	CreatedAt time.Time
	UpdatedAt time.Time
}

type Role struct {
	ID          uuid.UUID
	Name        string
	Description *string
	Permissions []string
	CreatedAt   time.Time
	UpdatedAt   time.Time
}

type Shift struct {
	ID         uuid.UUID
	EmployeeID uuid.UUID
	LocationID pgtype.UUID
	StartsAt   time.Time
	EndsAt     time.Time
	Notes      *string
	Status     string
	CreatedAt  time.Time
	UpdatedAt  time.Time
}